Promote NuGet detector experiment to primary detector
