+++ src/Microsoft.ComponentDetection.Orchestrator/Extensions/ServiceCollectionExtensions.cs
@@ -63,7 +63,6 @@ public static class ServiceCollectionExtensions
         services.AddSingleton<IExperimentProcessor, DefaultExperimentProcessor>();
         services.AddSingleton<IExperimentConfiguration, SimplePipExperiment>();
         services.AddSingleton<IExperimentConfiguration, RustCliDetectorExperiment>();
-        services.AddSingleton<IExperimentConfiguration, NuGetTargetFrameworkExperiment>();
 
         // Detectors
         // CocoaPods
@@ -105,7 +104,6 @@ public static class ServiceCollectionExtensions
         services.AddSingleton<IComponentDetector, NuGetComponentDetector>();
         services.AddSingleton<IComponentDetector, NuGetPackagesConfigDetector>();
         services.AddSingleton<IComponentDetector, NuGetProjectModelProjectCentricComponentDetector>();
-        services.AddSingleton<IComponentDetector, NuGetPackageReferenceFrameworkAwareDetector>();
 
         // PIP
         services.AddSingleton<IPyPiClient, PyPiClient>();
